Initial commit, this branch will serve as entry point fo the new components
