+++ app/page.tsx
@@ -1,25 +1,7 @@
-'use client'
-import { useState } from 'react'
-import AddItemModal from './_components/forms/add-item-modal'
-import ItemList from './_components/lists/items-list'
+import Cabinet from './_components/layouts/wardrobe-layout'
 
-type Item = {
-  name: string
-  type: 'top' | 'pants' | 'shoes'
-  image: string
-}
-
-export default function ItemManager() {
-  const [items, setItems] = useState<Item[]>([])
-
-  const addItem = (item: Item) => {
-    setItems([...items, item])
-  }
-
+export default function Wardrobe() {
   return (
-    <div className="container mx-auto p-4">
-      <AddItemModal addItem={addItem} />
-      <ItemList items={items} />
-    </div>
+<Cabinet/>
   )
 }
